taal configuraties verplaatst naar eigen bestanden voor duidelijkheid
