+++ frontend/frontend/src/i18n/config.ts
@@ -1,41 +1,8 @@
 import {initReactI18next} from "react-i18next";
 import i18n from "i18next";
 import LanguageDetector from "i18next-browser-languagedetector";
-
-const english = {
-    logo: "/assets/logo_UGent_EN_RGB_2400_white.png",
-    logo_blue: "/assets/logo_UGent_EN_RGB_2400_color.png",
-    login: "Log in with your UGent account",
-    back: "Back",
-    current_courses: "Current courses",
-    archived: "Archived",
-    students: "Students: ",
-    no_deadline: "No deadline",
-    add_admin: "Add admin",
-    submission: "Submission",
-    error: "Something went wrong.",
-    assignment: "Assignment:",
-    filename: "Submitted file:",
-    restrictions: "Restrictions:",
-    current_projects: "Current projects",
-};
-const dutch = {
-    logo: "/assets/logo_UGent_NL_RGB_2400_wit.png",
-    logo_blue: "/assets/logo_UGent_NL_RGB_2400_kleur.png",
-    login: "Log in met je UGent account",
-    back: "Terug",
-    current_courses: "Huidige vakken",
-    archived: "Gearchiveerd",
-    students: "Studenten: ",
-    no_deadline: "Geen deadline",
-    add_admin: "Voeg admin toe",
-    submission: "Indiening",
-    error: "Er is iets misgegaan.",
-    assignment: "Opgave:",
-    filename: "Ingediend bestand:",
-    restrictions: "Restricties:",
-    current_projects: "Huidige projecten",
-};
+import english from './en';
+import dutch from './nl';
 
 i18n.use(initReactI18next).use(LanguageDetector).init({
     fallbackLng: "en",
