+++ frontend/frontend/src/i18n/en.ts
@@ -1,0 +1,19 @@
+const english = {
+    logo: "/assets/logo_UGent_EN_RGB_2400_white.png",
+    logo_blue: "/assets/logo_UGent_EN_RGB_2400_color.png",
+    login: "Log in with your UGent account",
+    back: "Back",
+    current_courses: "Current courses",
+    archived: "Archived",
+    students: "Students: ",
+    no_deadline: "No deadline",
+    add_admin: "Add admin",
+    submission: "Submission",
+    error: "Something went wrong.",
+    assignment: "Assignment:",
+    filename: "Submitted file:",
+    restrictions: "Restrictions:",
+    current_projects: "Current projects",
+};
+
+export default english;
