+++ frontend/frontend/src/i18n/nl.ts
@@ -1,0 +1,19 @@
+const dutch = {
+    logo: "/assets/logo_UGent_NL_RGB_2400_wit.png",
+    logo_blue: "/assets/logo_UGent_NL_RGB_2400_kleur.png",
+    login: "Log in met je UGent account",
+    back: "Terug",
+    current_courses: "Huidige vakken",
+    archived: "Gearchiveerd",
+    students: "Studenten: ",
+    no_deadline: "Geen deadline",
+    add_admin: "Voeg admin toe",
+    submission: "Indiening",
+    error: "Er is iets misgegaan.",
+    assignment: "Opgave:",
+    filename: "Ingediend bestand:",
+    restrictions: "Restricties:",
+    current_projects: "Huidige projecten",
+};
+
+export default dutch;
